Fix minor issues and refactoring

+++ GpsTracker/Config/Constants.cs
@@ -6,7 +6,7 @@ namespace GpsTracker.Config
         public const int LocationUpdateInterval = 1000; //milliseconds
         public const int LocationUpdateFastestInterval = 1000; //milliseconds
         public const int DefaultMapZoom = 18;
-        public const double MinimalDisplacement = 1;//meters
+        public const int MinimalDisplacement = 1;//meters
         public const int FitTrackToScreenPadding = 70;
         public const string PolylineColor = "#AA3E97D1";
     }
